+++ Source/ACE.Server/Managers/EmoteManager.cs
@@ -1264,7 +1264,8 @@ namespace ACE.Server.Managers
             if (useRNG)
             {
                 var rng = ThreadSafeRandom.Next(0.0f, 1.0f);
-                emoteSet = emoteSet.Where(e => e.Probability >= rng);
+                emoteSet = emoteSet.OrderBy(e => e.Probability).Where(e => e.Probability >= rng);
+                //emoteSet = emoteSet.Where(e => e.Probability >= rng);
             }
 
             return emoteSet.FirstOrDefault();