+++ Source/ACE.Server/WorldObjects/WorldObject.cs
@@ -358,7 +358,7 @@ namespace ACE.Server.WorldObjects
             return isVisible;
         }
 
-        public bool IsDirectVisible(WorldObject wo, Position pos)
+        public bool IsDirectVisible(Position pos)
         {
             if (PhysicsObj == null)
                 return false;
@@ -1052,5 +1052,7 @@ namespace ACE.Server.WorldObjects
                     currentMotion = CurrentMotionState.MotionState.ForwardCommand;
             }
         }
+
+        public virtual bool IsAttunedOrContainsAttuned => (Attuned ?? 0) >= 1;
     }
 }